Specify service account for security API initContainer
When running the security API deployment with the optional initContainer "renew-sk-secret", that container needs to run under a privileged service account or else renew-sk-secret doesn't get created. I embedded the statement inside the conditional for the initContainer, so that the privileged service account isn't used unless needed.  I couldn't find a way to have the SA apply only to the initContainer.